WIP
* Updated upstream fork point in master (should already be merged, but that commit probably won't run, this will).
* The VK visual processor is selected by giving VKInteractiveVisualManagerFactory top position in the @ServiceProvider annotation
* Suppressed a bunch of warnings so the Core Display Vulkan issues are easier to spot.
* Added VK graph visual manager factory and made the function throw.  We need to decide how far up a ctor failure is propogated (we could handle it silently at the  point of failure but then we get left with broken instances).
* Full halt in the renderer currently causes a rendering deadlock.
* Minor update to Vector2f (ctor and getters).
* Added classifier to ivy retrieve pattern as without it the artifacts of the various LWJGL-* dependencies attempt to write to the same file.

+++ CoreGraphFramework/src/au/gov/asd/tac/constellation/graph/undo/UndoGraphEditOperation.java
@@ -716,14 +716,14 @@ public enum UndoGraphEditOperation {
         @Override
         public void execute(final UndoGraphEditState state, final GraphWriteMethods graph) {
             if (VERBOSE) {
-                System.out.println("setDoubleValueOperation.execute(" + state.getCurrentAttribute() + ", " + state.getCurrentId() + ", " + Double.longBitsToDouble((long) (Double.doubleToRawLongBits(graph.getDoubleValue(state.getCurrentAttribute(), state.getCurrentId())) ^ state.getCurrentDouble())) + ")");
+                System.out.println("setDoubleValueOperation.execute(" + state.getCurrentAttribute() + ", " + state.getCurrentId() + ", " + Double.longBitsToDouble((Double.doubleToRawLongBits(graph.getDoubleValue(state.getCurrentAttribute(), state.getCurrentId())) ^ state.getCurrentDouble())) + ")");
             }
-            graph.setDoubleValue(state.getCurrentAttribute(), state.getCurrentId(), Double.longBitsToDouble((long) (Double.doubleToRawLongBits(graph.getDoubleValue(state.getCurrentAttribute(), state.getCurrentId())) ^ state.getCurrentDouble())));
+            graph.setDoubleValue(state.getCurrentAttribute(), state.getCurrentId(), Double.longBitsToDouble((Double.doubleToRawLongBits(graph.getDoubleValue(state.getCurrentAttribute(), state.getCurrentId())) ^ state.getCurrentDouble())));
         }
 
         @Override
         public void undo(final UndoGraphEditState state, final GraphWriteMethods graph) {
-            graph.setDoubleValue(state.getCurrentAttribute(), state.getCurrentId(), Double.longBitsToDouble((long) (Double.doubleToRawLongBits(graph.getDoubleValue(state.getCurrentAttribute(), state.getCurrentId())) ^ state.getCurrentDouble())));
+            graph.setDoubleValue(state.getCurrentAttribute(), state.getCurrentId(), Double.longBitsToDouble((Double.doubleToRawLongBits(graph.getDoubleValue(state.getCurrentAttribute(), state.getCurrentId())) ^ state.getCurrentDouble())));
         }
 
         @Override
